+++ src/main/webapp/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html >
+<html>
+<head>
+    <title>Заголовок</title>
+</head>
+<body>
+"index.html - hello page!"
+</body>
+</html>
