--- src/main/webapp/index.jsp
@@ -1,5 +1,0 @@
-<html>
-<body>
-<h2>Hello World!привет</h2>
-</body>
-</html>
